Server: Add configurations for individual and server
Normal users don't need a postgres database to use this software.

Provide a single-user mode where a JSON is placed in the users
directory and it's read similar to the older versions of the server.

Note: Single-user mode is meant only for single users, using a database
is far more performant and addition of more users is accomplished
in the same way as single-user mode.
