+++ server/src/Config/credentialsHelper.ts
@@ -2,7 +2,13 @@ import { promises as fs } from 'fs';
 import { ProjectCredentials } from '../types';
 
 export async function handleCredentials(path: string): Promise<ProjectCredentials> {
-    const rawCreds = await fs.readFile(path, 'utf8');
+    const rawCreds = await fs
+        .readFile(path, 'utf8')
+        .catch((err) => console.log('No credentials found! Skipping...'));
+
+    if (!rawCreds) {
+        return {};
+    }
 
     const creds: ProjectCredentials = JSON.parse(rawCreds);
